windows: handle watch and event idiosyncrasies
Windows file creation events propagate as `CreateKind::Any` rather than
`CreateKind::File`. The same is true for `ModifyKind`, so both are
handled separately if `target_os` is windows.

Secondly, `Watcher::watch` is platform-specific, wherein calling it
twice for the same directory in windows will produce two identical
events messages for any singular event that is associated with the
directory. This does not occur on linux, but gating the call behind
`watch_exists` fixes the windows case, while maintaining the same
underlying behavior in other platforms.