reader: rework handling events to be asynchronous
Pushes the `*reader*` maps into an `Inner` structure which is
temporarily taken by async fn `handle_event` so that it can be polled in
the `impl Stream` as an anonymous Future.

This forgoes the original plan to implement the entire state machine by
hand, to prevent having to shovel and box memory on every iteration.
However, the `tokio::fs` operations introduce too many variants to the
state to be practical, so boxing the Future here is a reasonable
tradeoff.

Also, nuke `next_line_internal` in favor of just polling the reader
directly, to avoid dropping a Future on every iteration.