test: fix unit tests for osx
Switching from fsevents to kqueue broke some of the unit tests when run
on MacOS. The unit tests aren't a great source of confidence anyway,
since they so heavily rely on system-level operations. Integration and
in-situ testing is better for providing confidence.

This also re-enables osx unit testing in CI.